% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bpca.R
\name{bpca}
\alias{bpca}
\title{experimental basilisk approach to PCA}
\usage{
bpca(mat, ...)
}
\arguments{
\item{mat}{R matrix}

\item{\dots}{pass python-specific args to sklearn.decomposition.PCA}
}
\description{
experimental basilisk approach to PCA
}
\examples{
irismat = data.matrix(iris[,1:4])
bpca(irismat)
}
